storage_test_again/requirements.txt
psycopg2-binary==2.9.9
pymongo==4.6.0
python-dotenv==1.0.0
faker==19.6.2